+++ src/path_util.rs
@@ -3,7 +3,6 @@ use std::ffi::CString;
 use std::os::unix::ffi::OsStrExt;
 use std::path::{Path, PathBuf};
 use std::path::Component::RootDir;
-use std::fs::metadata;
 
 
 pub trait ToRelative {