+++ src/process_util.rs
@@ -1,8 +1,10 @@
 use std::env;
 use std::io::{self, Read};
 use std::path::{Path, PathBuf};
+use std::os::unix::io::FromRawFd;
 
 use libc::{getuid, c_int};
+use nix;
 use nix::sys::signal::{SIGINT, SIGTERM, SIGCHLD, SIGTTIN, SIGTTOU, SIGCONT};
 use nix::sys::signal::{SIGQUIT};
 use unshare::{Command, Stdio, Fd, ExitStatus, UidMap, GidMap, child_events};
@@ -24,6 +26,14 @@ pub static DEFAULT_PATH: &'static str =
 pub static PROXY_ENV_VARS: [&'static str; 5] =
     [ "http_proxy", "https_proxy", "ftp_proxy", "all_proxy", "no_proxy" ];
 
+
+pub fn squash_stdio(cmd: &mut Command) -> Result<(), String> {
+    let fd = try!(nix::unistd::dup(2)
+        .map_err(|e| format!("Can't duplicate fd 2: {}", e)));
+    cmd.stdout(unsafe { Stdio::from_raw_fd(fd) });
+    cmd.stdin(Stdio::null());
+    Ok(())
+}
 
 pub fn capture_stdout(mut cmd: Command) -> Result<Vec<u8>, String> {
     cmd.stdout(Stdio::piped());